Separate sort logic
