+++ lib/components/RunFrameWithApi/EnhancedFileSelectorCombobox/parseFilesToTree.ts
@@ -1,3 +1,5 @@
+import { naturalSort } from "./naturalSort"
+
 interface FileNode {
   name: string
   path: string
@@ -73,26 +75,6 @@ function getCurrentFolderContents(
 
   const files = targetNode.children.filter((node) => node.type === "file")
   const folders = targetNode.children.filter((node) => node.type === "folder")
-
-  // Natural sort comparator for numeric parts
-  const naturalSort = (a: string, b: string) => {
-    const extA = a.lastIndexOf(".")
-    const extB = b.lastIndexOf(".")
-    const baseA = extA > 0 ? a.substring(0, extA) : a
-    const baseB = extB > 0 ? b.substring(0, extB) : b
-
-    // This ensures "Foo" comes before "Foo_bar" (both with same extension)
-    if (baseA !== baseB && a.substring(extA) === b.substring(extB)) {
-      if (baseA.startsWith(baseB)) {
-        return 1
-      }
-      if (baseB.startsWith(baseA)) {
-        return -1
-      }
-    }
-
-    return a.localeCompare(b, undefined, { numeric: true, sensitivity: "base" })
-  }
 
   // Sort files first, then folders, using natural sorting
   files.sort((a, b) => naturalSort(a.name, b.name))
